RequestHeader, ResponseBuilder 역할 분리 v1
